requirements.txt
aiofile==3.5.0
aiohttp==3.7.4.post0
async-timeout==3.0.1
asyncio==3.4.3
attrs==21.2.0
caio==0.8.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
idna==2.10
multidict==5.1.0
protobuf==3.17.1
pycparser==2.20
python-axolotl==0.2.3
python-axolotl-curve25519==0.4.1.post2
requests==2.25.1
six==1.16.0
typing-extensions==3.10.0.0
urllib3==1.26.5
yarl==1.6.3
